+++ crawler/updater/alma.py
@@ -9,7 +9,8 @@ from crawler.web.generic import url_get_last_modified
 from crawler.web.directory import web_get_checksum, web_get_current_image_metadata
 
 from bs4 import BeautifulSoup
-from pprint import pprint
+from loguru import logger
+
 
 def build_image_url(release, versionpath):
     if not release["baseURL"].endswith("/"):
@@ -70,6 +71,10 @@ def get_metadata(release, image_filedate):
             if release_date is None:
                 return None
 
+            logger.debug("url: " + build_image_url(release, release_version_path))
+            logger.debug("last version: " + version)
+            logger.debug("release_date: " + release_date)
+
             return {
                 "url": build_image_url(
                     release, release_version_path
@@ -122,22 +127,22 @@ def alma_update_check(release, last_checksum):
 
     checksum_url = base_url + release["releasepath"] + "/" + release["checksumname"]
 
-    print("alma_update_check/checksum_url:" + checksum_url)
+    logger.debug("checksum_url: " + checksum_url)
 
     # as specified in image-sources.yaml
     # imagename: AlmaLinux-9-GenericCloud-latest.x86_64
     # extension: qcow2
     imagename = release["imagename"] + "." + release["extension"]
 
-    print("alma_update_check/imagename:" + imagename)
+    logger.debug("imagename: " + imagename)
 
     current_checksum = web_get_checksum(checksum_url, imagename)
 
-    print("alma_update_check/current_checksum:" + current_checksum)
+    logger.debug("current_checksum: " + current_checksum)
 
     if current_checksum is None:
-        print(
-            "ERROR: no matching checksum found - check image (%s) "
+        logger.error(
+            "no matching checksum found - check image (%s) "
             "and checksum filename (%s)" % (imagename, release["checksumname"])
         )
         return None
@@ -147,18 +152,18 @@ def alma_update_check(release, last_checksum):
     current_checksum = release["algorithm"] + ":" + current_checksum
 
     if current_checksum != last_checksum:
-        print("DO something")
+        logger.debug("current_checksum " + current_checksum + " differs from last_checksum " + last_checksum)
         image_url = base_url + release["releasepath"] + "/" + imagename
 
-        print("alma_update_check/image_url:" + image_url)
+        logger.debug("image_url:" + image_url)
 
         image_filedate = url_get_last_modified(image_url)
 
-        print("alma_update_check/image_filedate:" + image_filedate)
+        logger.debug("image_filedate:" + image_filedate)
 
         image_metadata = get_metadata(release, image_filedate)
         if image_metadata is not None:
-            pprint(image_metadata)
+            logger.debug("got metadata")
             update = {}
             update["release_date"] = image_metadata["release_date"]
             update["url"] = image_metadata["url"]
@@ -166,6 +171,7 @@ def alma_update_check(release, last_checksum):
             update["checksum"] = current_checksum
             return update
         else:
+            logger.warning("got no metadata")
             return None
 
     return None