+++ image-crawler.py
@@ -11,7 +11,8 @@
 import argparse
 import sys
 import os
-
+# import logging
+from loguru import logger
 from crawler.core.config import config_read
 from crawler.core.database import (
     database_connect,
@@ -22,9 +23,14 @@ from crawler.core.exporter import export_image_catalog, export_image_catalog_all
 from crawler.core.main import crawl_image_sources
 from crawler.git.base import clone_or_pull, update_repository
 
+# logging.basicConfig(level=logging.DEBUG, format='%(asctime)s %(name)s %(levelname)s:%(message)s')
+
+# logging.basicConfig(level=logging.DEBUG)
+# logger = logging.getLogger(__name__)
+
 
 def main():
-    print("\nplusserver Image Crawler v0.4.0\n")
+    logger.info("plusserver Image Crawler v0.4.0 started")
 
     working_directory = os.getcwd()
     program_directory = os.path.dirname(os.path.abspath(__file__))
@@ -74,7 +80,8 @@ def main():
 
     config = config_read(config_filename, "configuration")
     if config is None:
-        raise SystemExit("\nERROR: Unable to open config " + config_filename)
+        logger.error("ERROR: Unable to open config " + config_filename)
+        raise SystemExit(1)
 
     # read the image sources
     if args.sources is not None:
@@ -108,28 +115,28 @@ def main():
             git_ssh_command,
         )
     else:
-        print("No image catalog repository configured")
+        logger.warning("No image catalog repository configured")
 
     # connect to database
     database = database_connect(config["database_name"])
     if database is None:
-        print("\nERROR: Could not open database %s" % config["database_name"])
-        print(
-            '\nRun "./image-crawler.py --init-db" to create a new database OR config check your etc/config.yaml'
+        logger.error("Could not open database %s" % config["database_name"])
+        logger.error(
+            'Run "./image-crawler.py --init-db" to create a new database OR config check your etc/config.yaml'
         )
         sys.exit(1)
 
     # crawl image sources when requested
     if args.export_only:
-        print("\nSkipping repository crawling")
+        logger.info("Skipping repository crawling")
         updated_sources = {}
     else:
-        print("\nStart repository crawling")
+        logger.info("Start repository crawling")
         updated_sources = crawl_image_sources(image_source_catalog, database)
 
     # export image catalog
     if args.updates_only:
-        print("\nSkipping catalog export")
+        logger.info("Skipping catalog export")
     else:
         if config["local_repository"].startswith("/"):
             export_path = config["local_repository"]
@@ -137,7 +144,7 @@ def main():
             export_path = working_directory + "/" + config["local_repository"]
 
         if updated_sources:
-            print("\nExporting catalog to %s" % export_path)
+            logger.info("Exporting catalog to %s" % export_path)
             export_image_catalog(
                 database,
                 image_source_catalog,
@@ -147,7 +154,7 @@ def main():
             )
         else:
             if args.export_only:
-                print("\nExporting all catalog files to %s" % export_path)
+                logger.info("Exporting all catalog files to %s" % export_path)
                 export_image_catalog_all(
                     database,
                     image_source_catalog,
@@ -161,10 +168,11 @@ def main():
             database, config["local_repository"], updated_sources, git_ssh_command
         )
     else:
-        print("No remote repository update needed.")
+        logger.info("No remote repository update needed.")
 
     database_disconnect(database)
 
 
 if __name__ == "__main__":
+
     main()